feat(home): consolidate three dashboards into single home view
Merge Operations, Membership, and Owner dashboards into one
unified /app/ page. Single Promise.all with 13 queries covering
revenue, active members, ARR, renewals, alerts, today's schedule,
and upcoming renewals table. Tech role still redirects to My Day.
Redirect /app/operations and /app/membership-dashboard to /app/.
